requirements.txt
# --- Alapvető AI motorok ---
exllamav2
# A GGUF slotokhoz (Llama-cpp-python) - Ez kritikus a King/Scribe/Valet számára!
llama-cpp-python

# --- Backend és API ---
fastapi
uvicorn[standard]
pyyaml
python-dotenv
pydantic

# --- RAG és Embedding (Ez kell az automata letöltéshez) ---
sentence-transformers
torch
transformers
accelerate
protobuf          # A Gemma tokenizálóhoz kötelező

# --- Adatbázis és Tárolás ---
qdrant-client
redis
networkx
numpy

# --- Monitorozás és Eszközök ---
colorama
huggingface_hub
psutil            # A monitor.py-hoz, hogy lásd a VRAM/CPU használatot
asyncio           # Az Orchestrator aszinkronitásához